requirements.txt
requests_html
requests
PyExecJS
Pillow
numpy